requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
psycopg2-binary==2.9.10
python-dotenv==1.0.0
openai==1.12.0
Werkzeug==3.0.1
WTForms==3.1.1
email-validator==2.1.0
gunicorn==21.1.0
eventlet==0.35.1
Pillow
